requirements.txt
discord.py
python-dotenv
beautifulsoup4
requests
lxml
pytrends
urllib3
praw
asyncpraw
spacy
#python -m spacy download en_core_web_sm